Métodos de sync movidos para uma classe própria
-- Os métodos syncNewTransaction, syncUpdateReceived e syncUpdateSent foram movidos para uma classe própria chamada 'sync'
